pyproject.toml
[tool.poetry]
name = "gitlab_ldap_group_sync"
version = "0.1.0"
description = "Sync Gitlab groups and ssh keys with ldap"
authors = ["Maximilian Nöthe <maximilian.noethe@tu-dortmund.de>"]
license = "mit"

[tool.poetry.dependencies]
python = "^3.6"
python-gitlab = "^2.1.2"
python-dotenv = "^0.12.0"
python-ldap = "^3.2.0"

[tool.poetry.dev-dependencies]
ipython = "^7.13.0"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
